commit new methods

+++ src/main/java/pro/sky/HomeWorkEmployeeMockito/api/EmployeeInterface.java
@@ -2,16 +2,18 @@ package pro.sky.HomeWorkEmployeeMockito.api;
 
 import pro.sky.HomeWorkEmployeeMockito.model.Employee;
 
+import java.util.Arrays;
+import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 
 public interface EmployeeInterface {
 
-    Employee add(String firstName, String lastName, int department, double salary);
+    Employee add(String firstName, String lastName, int department, int salary);
 
-    Employee remove(String firstName, String lastName, int department, double salary);
+    Employee remove(String firstName, String lastName, int department, int salary);
 
-    Employee find(String firstName, String lastName, int department, double salary);
+    Employee find(String firstName, String lastName, int department, int salary);
 
     Map<String, Employee> findAll();
 
@@ -20,4 +22,5 @@ public interface EmployeeInterface {
     public List<Employee> getEmployeesList();
 
     List<Employee> myList(Map<String, Employee> employeeMap);
+
 }
